LayoutTests/fast/attachment/attachment-select-on-click-expected.txt

| "text before "
| <#selection-anchor>
| <attachment>
|   id="attachment"
| <#selection-focus>
| " text after"
